Add delete and download

+++ types/user.go
@@ -1,7 +1,7 @@
 package types
 
 type User struct {
-	ID           uint64
+	ID           int32
 	Username     string
 	PasswordHash []byte
 	PasswordSalt []byte
